pyproject.toml
[tool.poetry]
name = "langchaintest"
version = "0.1.0"
description = ""
authors = ["Andrew Samaha <andrew.samaha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
langchain = "^0.0.308"
openai = "^0.28.1"
tiktoken = "^0.5.1"
faiss-cpu = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
